문자기능 수정 및 2차 수정 배포

+++ src/modules/consult-baemin-report/consult-baemin-report.entity.ts
@@ -28,7 +28,7 @@ export class ConsultBaeminReport extends BaseEntity<ConsultBaeminReport> {
   //   평점
   @Column({
     name: 'average_score',
-    type: 'int',
+    type: 'double',
   })
   averageScore: number;
